requirements.txt
discord.py==2.3.2
yt-dlp
PyNaCl
python-dotenv==1.0.0
